made the normal main a generic function as well
