+++ lib/src/workload.rs
@@ -5,6 +5,6 @@ pub use core_workload::CoreWorkload;
 use crate::db::DB;
 
 pub trait Workload {
-    fn do_insert<T: DB>(&self, db: T);
-    fn do_transaction<T: DB>(&self, db: T);
+    fn do_insert<T: DB>(&self, db: &T);
+    fn do_transaction<T: DB>(&self, db: &T);
 }
